Fix benchmarks running indefinitely
It turns out that 'seq' doesn't work the way I thought it worked.

https://ircbrowse.tomsmeding.com/day/lchaskell/2024/02/02?id=1203227#trid1203227

The computation
> a `seq` b
has the same value as the computation
> b
but is strict in 'a'. This does _not_ mean that 'a' will have been
reduced to WHNF when 'a `seq` b' has. This means 'seq' (and by
extension, 'deepseq') is unsuitable if you want to ensure that a
particular value is fully evaluated before continuing in a monadic
program -- which is what I was trying to do.

Instead, one should use Control.Exception.evaluate, which does _not_ use
'seq' but instead a primop called 'seq#'. Some history can be found
here:

https://gitlab.haskell.org/ghc/ghc/-/issues/2273

I believe 'seq#' implements the idea of SPJ's 'fseq', which works in IO
because IO is a state monad, i.e. it has a continuation right there.