Run first job outside of thread pool
Due to the CPSing this is allowed (all work after the first fork will
run in the thread pool, meaning that there is never work outside and
inside the thread pool running in parallel), and it makes sequential
benchmarks _much_ faster due to avoiding the thread pool overhead.

There is of course still the overhead of the more complicated monad, but
the thread pool (sometimes? at least for rotvecquat?) has a much bigger
effect (~40ms).